pyproject.toml
[tool.poetry]
name = "FastCLAM"
version = "0.0.1"
description = "FastAPI for CLAMS Applications"
authors = ["WGBH-MLA"]
readme = "README.md"
packages = [{include = "fastclam"}]

[tool.poetry.dependencies]
python = "^3.8.1"
clams-python = "^0.5.2"
mmif-python = "^0.4.2"
pydantic = "^1.8.2"
fastapi = "^0.94.0"
loguru = "^0.6.0"
rich = "^13.3.2"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.21.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.2.2"
httpx = "^0.23.3"
pytest-cov = "^4.0.0"
black = "^23.1.0"
flake8 = "^6.0.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.1.2"
mkdocs-git-revision-date-localized-plugin = "^1.2.0"
mkdocstrings = {extras = ["python"], version = "^0.20.0"}
mike = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
